(PC-40065)[PRO] refactor: refactoring css / change wording variable / type
