--- pro/src/pages/Homepage/components/CollectiveOffersBookableCard/CollectiveOffersBookableCard.module.scss
@@ -1,24 +1,0 @@
-@use "styles/mixins/_size.scss" as size;
-
-/**
- * We setup mobile display by default
- *   (cf the following min-width in media queries)
- *
- * It allows us to limit overrides and code complexity.
- *
- * @see ./components/CollectiveOffersBookableLine/CollectiveOffersBookableLine.module.scss
- */
-.offer-list {
-  // We prefer grid to flex here to be consistent with tablet and desktop displays
-  display: grid;
-  grid-template-columns: 1fr;
-  gap: var(--size-spacing-l);
-
-  @media screen and (min-width: size.$laptop) {
-    grid-template-columns: max-content 1fr max-content;
-  }
-
-  @media screen and (min-width: size.$desktop) {
-    grid-template-columns: max-content 1fr auto max-content;
-  }
-}
