+++ pro/src/pages/Homepage/components/CollectiveOffersLine.module.scss
@@ -1,4 +1,3 @@
-@use "styles/mixins/_rem.scss" as rem;
 @use "styles/mixins/_size.scss" as size;
 @use "styles/mixins/_fonts.scss" as fonts;
 @use "styles/mixins/_a11y.scss" as a11y;
@@ -9,7 +8,7 @@
  *
  * It allows us to limit overrides and code complexity.
  *
- * @see ../../CollectiveOffersBookableCard/CollectiveOffersBookableCard.module.scss
+ * @see ./CollectiveOffersList.module.scss
  */
 .offer-line {
   display: grid; // In mobile, every line has its own grid
