+++ pro/src/pages/Homepage/components/CollectiveOffersList.module.scss
@@ -6,7 +6,7 @@
  *
  * It allows us to limit overrides and code complexity.
  *
- * @see ./components/CollectiveOffersTemplateLine/CollectiveOffersTemplateLine.module.scss
+ * @see ./CollectiveOffersLine.module.scss
  */
 .offer-list {
   // We prefer grid to flex here to be consistent with tablet and desktop displays
